Prepare for deploy

+++ school-management-system/src/components/Screens/fees.jsx
@@ -35,12 +35,15 @@ const handlePrint = useReactToPrint({
   content: () => componentRef.current,
 });
 const removeData = async(id) => {
+  let flag= window.confirm("Delete  record!")
+  if(flag)
+  { 
     await axios.delete(`/api/fee`, { params: {id} }) 
         .then(res => {
             const del = gdata.filter(gdata => id !== gdata._id)
             setData(del)
            
-        }) 
+        }) }
 }
 
 useEffect(()=>{
@@ -89,7 +92,7 @@ const generateDefaultFee = async()=>{
 }
 const deleteRecord = async()=>{
   if(month && year ){
-  let flag= window.confirm("Delete Paid record of specific month")
+  let flag= window.confirm("Delete  record of specific month")
   if(flag)
   { 
     await axios.delete('/api/deleteFee', { params: {month,year} })
